Moved domain classes to domain package
